+++ saby_combat/templates/base.html
@@ -46,7 +46,7 @@
                 </div>
                 <div class="container-for-element-menu">
                     <div class="side-decoration"></div>
-                    <a class="block" href="{{ url_for('friends_page') }}">
+                    <a class="block" href="{{ url_for('community') }}">
                         <div class="block-for-img">
                             <img src="{{ url_for('static', filename='images-utils/images-sidebar/sidebar_comminuty.svg') }}" alt="Изображение вкладки сообщество">
                         </div>
